Added in SQlite database for mocking transactions

+++ app/src/main/java/com/example/financialspendingdashboardanalysis/model/FinancialTransactionData.kt
@@ -1,9 +1,11 @@
 package com.example.financialspendingdashboardanalysis.model
 
+import com.example.financialmodels.TransactionCategory
+
 data class FinancialTransactionData(
     val transactionCategory: TransactionCategory = TransactionCategory.UNKNOWN,
     val transactionType: String = "",
-    val amount: Int = 0,
+    val amount: Long = 0,
     val fromAccount: String = "",
     val fundAccount: String = "",
     val paymentType: String = "",
