+++ app/src/main/java/com/example/financialspendingdashboardanalysis/ui/theme/DisplayFinancialPieChart.kt
@@ -1,10 +1,8 @@
 package com.example.financialspendingdashboardanalysis.ui.theme
 
 import android.graphics.Paint
-import android.util.Log
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -21,7 +19,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -33,11 +30,9 @@ import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.graphics.drawscope.scale
 import androidx.compose.ui.graphics.nativeCanvas
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.unit.dp
-import com.example.financialspendingdashboardanalysis.model.BarGraphInfo
 import com.example.financialspendingdashboardanalysis.model.PieChartInfo
-import com.example.financialspendingdashboardanalysis.model.TransactionCategory
+import com.example.financialmodels.TransactionCategory
 import kotlin.math.atan2
 import kotlin.math.cos
 import kotlin.math.min
@@ -98,7 +93,7 @@ fun DisplayFinancialPieChart(
     val categoryNames: MutableList<String> = remember { mutableListOf() }
     var pieCenter: Offset = Offset.Zero
     var diameter: Float = 0.toFloat()
-    var overallAmountForAllCategory = 0
+    var overallAmountForAllCategory = 0L
 
     LaunchedEffect(Unit) {
         animationPlayed.value = true
@@ -306,7 +301,7 @@ fun DisplayFinancialPieChart(
 @Composable
 fun PieChartLabelIndexDisplayed(
     generateRandomColorsList: List<Color>,
-    getMonthlyPieData: Map<String, Int>
+    getMonthlyPieData: Map<String, Long>
 ) {
     Column(modifier = Modifier.fillMaxWidth()) {
         getMonthlyPieData.entries.forEachIndexed { index, (key, entry) ->
@@ -326,7 +321,7 @@ fun PieChartLabelIndexDisplayed(
  * Time Complexity: O(1)
  * Space Complexity: O(1)
  */
-fun convertValueToAMount(inputAmount: Int): String {
+fun convertValueToAMount(inputAmount: Long): String {
     val formatter = java.text.DecimalFormat("#,##0.00").apply {
         decimalFormatSymbols = decimalFormatSymbols.apply {
             groupingSeparator = ' '
